CIV-1805 : Light Refactor of tech debt
