+++ src/main/java/uk/gov/hmcts/cmc/claimstore/models/idam/User.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.claimstore.idam.models;
+package uk.gov.hmcts.cmc.claimstore.models.idam;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
