+++ src/main/java/uk/gov/hmcts/cmc/claimstore/models/idam/UserDetails.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.claimstore.idam.models;
+package uk.gov.hmcts.cmc.claimstore.models.idam;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
