--- src/test/java/uk/gov/hmcts/cmc/claimstore/courtfinder/models/AddressTest.java
@@ -1,22 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.courtfinder.models;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-class AddressTest {
-
-    @Test
-    void shouldStripAddressFromSpecialChars() {
-        Address address = Address.builder()
-            .addressLines(Arrays.asList("aaa\r", "bbb\r"))
-            .postcode("EC2Y3DD")
-            .town("London")
-            .type("Visit or contact us")
-            .build();
-        assertThat(address.toString()).isEqualTo("aaa\nbbb\nEC2Y3DD\nLondon");
-    }
-
-}
